[C API] Rename include directory svs/c_api/ to svs/c/
