+++ bindings/c/docs/C_API_Design.md
@@ -413,9 +413,9 @@ function returns an empty result set for that query.
 ### Public Headers
 
 ```c
-#include "svs/c_api/svs_c.h"           // Main C API
-#include "svs/c_api/svs_c_config.h"    // API configuration macroses
-#include "svs/c_api/svs_c_version.h"   // SVS_C_API_VERSION[_MAJOR|_MINOR|_PATCH|_STRING]
+#include "svs/c/svs_c.h"           // Main C API
+#include "svs/c/svs_c_config.h"    // API configuration macroses
+#include "svs/c/svs_c_version.h"   // SVS_C_API_VERSION[_MAJOR|_MINOR|_PATCH|_STRING]
 ```
 
 `svs_c_version.h` is generated by CMake from `svs_c_version.h.in` and defines both
