+++ bindings/c/include/svs/c/svs_c_config.h
@@ -16,7 +16,7 @@
 
 #pragma once
 
-#include "svs/c_api/svs_c_version.h"
+#include "svs/c/svs_c_version.h"
 
 // All symbols shall be internal unless marked as SVS_API
 #if defined _WIN32 || defined __CYGWIN__
