+++ bindings/c/src/algorithm.hpp
@@ -15,7 +15,7 @@
  */
 #pragma once
 
-#include "svs/c_api/svs_c.h"
+#include "svs/c/svs_c.h"
 
 // #include <svs/concepts/data.h>
 // #include <svs/core/distance.h>
