+++ bindings/c/src/data_builder/simple.hpp
@@ -15,7 +15,7 @@
  */
 #pragma once
 
-#include "svs/c_api/svs_c.h"
+#include "svs/c/svs_c.h"
 
 #include "allocator.hpp"
 #include "storage.hpp"
