+++ bindings/c/src/error.cpp
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "svs/c_api/svs_c.h"
+#include "svs/c/svs_c.h"
 
 #include "error.hpp"
