+++ bindings/c/src/error.hpp
@@ -15,7 +15,7 @@
  */
 #pragma once
 
-#include "svs/c_api/svs_c.h"
+#include "svs/c/svs_c.h"
 
 #include <stdexcept>
 #include <string>
