+++ bindings/c/src/index.hpp
@@ -15,7 +15,7 @@
  */
 #pragma once
 
-#include "svs/c_api/svs_c.h"
+#include "svs/c/svs_c.h"
 
 #include "algorithm.hpp"
 #include "filtered_search.hpp"
