+++ bindings/c/src/threadpool.hpp
@@ -15,7 +15,7 @@
  */
 #pragma once
 
-#include "svs/c_api/svs_c.h"
+#include "svs/c/svs_c.h"
 
 #include "error.hpp"
 #include "types_support.hpp"
